[ADD] added playSound method, updated Items

+++ app/src/main/java/com/example/projectsnakereloaded/SpeedLoss.java
@@ -4,18 +4,18 @@ import processing.core.PImage;
 
 public class SpeedLoss extends Item{
 
-    private int speedLoss;
+    private double speedLoss;
 
     public SpeedLoss(int x, int y, PImage image) {
         super(x, y, image);
         speedLoss = 3;
     }
 
-    public int getSpeedLoss() {
+    public double getSpeedLoss() {
         return speedLoss;
     }
 
-    public void setSpeedLoss(int speedLoss) {
+    public void setSpeedLoss(double speedLoss) {
         this.speedLoss = speedLoss;
     }
 }
